Updated Visual Novel Tracker
Visual Novel tracker now includes player dialog options and a last selected dialog.
dialog is not split between dialogRendered, which shows the state of progressive disclosure and is reported in twees, and dialogFull, which is the complete dialog passage that will be rendered for a given step of dialog.
speaker emotion is now also displayed in the key frame corresponding to which player sprite is being used at the time.